+++ quick_blue/test/test_support/fake_quick_blue_platform.dart
@@ -60,7 +60,7 @@ class FakeQuickBluePlatform extends QuickBluePlatform {
   }
 
   @override
-  Stream<BlueBluetoothState> get bluetoothStateStream {
+  Stream<BlueBluetoothState> get bluetoothStateEvents {
     calls.add('bluetoothStateStream');
     if (!emitInitialBluetoothState) {
       return _bluetoothStateController.stream;